feat: common `getStaticDocumentProps`
